Control.Exception.Compat

Documentation

module Control.Exception

throw :: forall (r :: RuntimeRep) (a :: TYPE r) e. Exception e => e -> a

Throw an exception. Exceptions may be thrown from purely functional code, but may only be caught within the IO monad.

WARNING: You may want to use throwIO instead so that your pure code stays exception-free.