Data.Monoid.Compat

Documentation

newtype Any

Boolean monoid under disjunction (||).

Any x <> Any y = Any (x || y)

Examples

Expand

>>> Any True <> mempty <> Any False
Any {getAny = True}

>>> mconcat (map (\x -> Any (even x)) [2,4,6,7,8])
Any {getAny = True}

>>> Any False <> mempty
Any {getAny = False}

Constructors

Any

Fields

getAny :: Bool

newtype Sum a

Monoid under addition.

Sum a <> Sum b = Sum (a + b)

Examples

Expand

>>> Sum 1 <> Sum 2 <> mempty
Sum {getSum = 3}

>>> mconcat [ Sum n | n <- [3 .. 9]]
Sum {getSum = 42}

Constructors

Sum

Fields

getSum :: a

newtype Product a

Monoid under multiplication.

Product x <> Product y == Product (x * y)

Examples

Expand

>>> Product 3 <> Product 4 <> mempty
Product {getProduct = 12}

>>> mconcat [ Product n | n <- [2 .. 10]]
Product {getProduct = 3628800}

Constructors

Product

Fields

getProduct :: a

newtype Last a

Maybe monoid returning the rightmost non-Nothing value.

Last a is isomorphic to Dual (First a), and thus to Dual (Alt Maybe a)

Data.Semigroup.Last. The former returns the last non-Nothing, so x <> Data.Monoid.Last Nothing = x. The latter simply returns the last value, thus x <> Data.Semigroup.Last Nothing = Data.Semigroup.Last Nothing.

Examples

Expand

>>> Last (Just "hello") <> Last Nothing <> Last (Just "world")
Last {getLast = Just "world"}

>>> Last Nothing <> mempty
Last {getLast = Nothing}

Constructors

Last

Fields

getLast :: Maybe a

newtype First a

Maybe monoid returning the leftmost non-Nothing value.

First a is isomorphic to Alt Maybe a, but precedes it historically.

Beware that Data.Monoid.First is different from Data.Semigroup.First. The former returns the first non-Nothing, so Data.Monoid.First Nothing <> x = x. The latter simply returns the first value, thus Data.Semigroup.First Nothing <> x = Data.Semigroup.First Nothing.

Examples

Expand

>>> First (Just "hello") <> First Nothing <> First (Just "world")
First {getFirst = Just "hello"}

>>> First Nothing <> mempty
First {getFirst = Nothing}

Constructors

First

Fields

getFirst :: Maybe a

class Semigroup a => Monoid a where

The class of monoids (types with an associative binary operation that has an identity). Instances should satisfy the following:

Right identity

x <> mempty = x

Left identity

mempty <> x = x

Associativity

x <> (y <> z) = (x <> y) <> z (Semigroup law)

Concatenation

mconcat = foldr (<>) mempty

You can alternatively define mconcat instead of mempty, in which case the laws are:

Unit

mconcat (pure x) = x

Multiplication

mconcat (join xss) = mconcat (fmap mconcat xss)

Subclass

mconcat (toList xs) = sconcat xs

The method names refer to the monoid of lists under concatenation, but there are many other instances.

Some types can be viewed as a monoid in more than one way, e.g. both addition and multiplication on numbers. In such cases we often define newtypes and make those instances of Monoid, e.g. Sum and Product.

NOTE: Semigroup is a superclass of Monoid since base-4.11.0.0.

Minimal complete definition

mempty | mconcat

Methods

mempty :: a

Identity of mappend

Examples

Expand

>>> "Hello world" <> mempty
"Hello world"

>>> mempty <> [1, 2, 3]
[1,2,3]

mappend :: a -> a -> a

An associative operation

NOTE: This method is redundant and has the default implementation mappend = (<>) since base-4.11.0.0. Should it be implemented manually, since mappend is a synonym for (<>), it is expected that the two functions are defined the same way. In a future GHC release mappend will be removed from Monoid.

mconcat :: [a] -> a

Fold a list using the monoid.

For most types, the default definition for mconcat will be used, but the function is included in the class definition so that an optimized version can be provided for specific types.

>>> mconcat ["Hello", " ", "Haskell", "!"]
"Hello Haskell!"

newtype Alt (f :: k -> Type) (a :: k)

Monoid under <|>.

Alt l <> Alt r == Alt (l <|> r)

Examples

Expand

>>> Alt (Just 12) <> Alt (Just 24)
Alt {getAlt = Just 12}

>>> Alt Nothing <> Alt (Just 24)
Alt {getAlt = Just 24}

Since: base-4.8.0.0

Constructors

Alt

Fields

getAlt :: f a

newtype All

Boolean monoid under conjunction (&&).

All x <> All y = All (x && y)

Examples

Expand

>>> All True <> mempty <> All False)
All {getAll = False}

>>> mconcat (map (\x -> All (even x)) [2,4,6,7,8])
All {getAll = False}

>>> All True <> mempty
All {getAll = True}

Constructors

All

Fields

getAll :: Bool

newtype Endo a

The monoid of endomorphisms under composition.

Endo f <> Endo g == Endo (f . g)

Examples

Expand

>>> let computation = Endo ("Hello, " ++) <> Endo (++ "!")
>>> appEndo computation "Haskell"
"Hello, Haskell!"

>>> let computation = Endo (*3) <> Endo (+1)
>>> appEndo computation 1
6

Constructors

Endo

Fields

appEndo :: a -> a

newtype Dual a

The dual of a Monoid, obtained by swapping the arguments of mappend. | The dual of a Monoid, obtained by swapping the arguments of (<>).

Dual a <> Dual b == Dual (b <> a)

Examples

Expand

>>> Dual "Hello" <> Dual "World"
Dual {getDual = "WorldHello"}

>>> Dual (Dual "Hello") <> Dual (Dual "World")
Dual {getDual = Dual {getDual = "HelloWorld"}}

Constructors

Dual

Fields

getDual :: a

newtype Ap (f :: k -> Type) (a :: k)

This data type witnesses the lifting of a Monoid into an Applicative pointwise.

Examples

Expand

>>> Ap (Just [1, 2, 3]) <> Ap Nothing
Ap {getAp = Nothing}

>>> Ap [Sum 10, Sum 20] <> Ap [Sum 1, Sum 2]
Ap {getAp = [Sum {getSum = 11},Sum {getSum = 12},Sum {getSum = 21},Sum {getSum = 22}]}

Since: base-4.12.0.0

Constructors

Ap

Fields

getAp :: f a

(<>) :: Semigroup a => a -> a -> a infixr 6

An associative operation.

Examples

Expand

>>> [1,2,3] <> [4,5,6]
[1,2,3,4,5,6]

>>> Just [1, 2, 3] <> Just [4, 5, 6]
Just [1,2,3,4,5,6]

>>> putStr "Hello, " <> putStrLn "World!"
Hello, World!