System.IO.Compat

Documentation

module System.IO

getContents' :: IO String

The getContents' operation returns all user input as a single string, which is fully read before being returned (same as hGetContents' stdin).

Since: base-4.15.0.0

hGetContents' :: Handle -> IO String

The hGetContents' operation reads all input on the given handle before returning it as a String and closing the handle.

Since: base-4.15.0.0

readFile' :: FilePath -> IO String

The readFile' function reads a file and returns the contents of the file as a string. The file is fully read before being returned, as with getContents'.

Since: base-4.15.0.0