System.IO.Error.Compat

Documentation

module System.IO.Error

isResourceVanishedError :: IOError -> Bool

An error indicating that the operation failed because the resource vanished. See resourceVanishedErrorType.

Since: base-4.14.0.0

resourceVanishedErrorType :: IOErrorType

I/O error where the operation failed because the resource vanished. This happens when, for example, attempting to write to a closed socket or attempting to write to a named pipe that was deleted.

Since: base-4.14.0.0

isResourceVanishedErrorType :: IOErrorType -> Bool

I/O error where the operation failed because the resource vanished. See resourceVanishedErrorType.

Since: base-4.14.0.0