Data.Function.Compat

Documentation

module Data.Function

(&) :: a -> (a -> b) -> b infixl 1

& is a reverse application operator. This provides notational convenience. Its precedence is one higher than that of the forward application operator $, which allows & to be nested in $.

>>> 5 & (+1) & show
"6"

Since: base-4.8.0.0

applyWhen :: Bool -> (a -> a) -> a -> a

applyWhen applies a function to a value if a condition is true, otherwise, it returns the value unchanged.

It is equivalent to flip (bool id).

Algebraic properties:

applyWhen True = id

applyWhen False f = id

Since: base-4.18.0.0