Data.Proxy.Compat

Documentation

asProxyTypeOf :: a -> proxy a -> a

asProxyTypeOf is a type-restricted version of const. It is usually used as an infix operator, and its typing forces its first argument (which is usually overloaded) to have the same type as the tag of the second.