Data.Semigroup.Compat

Description

This backports the modern Data.Semigroup interface back to base-4.9/GHC 8.0.

Documentation

class Semigroup a where

The class of semigroups (types with an associative binary operation).

Since: 4.9.0.0

Methods

(<>) :: a -> a -> a infixr 6

An associative operation.

(a <> b) <> c = a <> (b <> c)

If a is also a Monoid we further require

(<>) = mappend

sconcat :: NonEmpty a -> a

Reduce a non-empty list with <>

The default definition should be sufficient, but this can be overridden for efficiency.

stimes :: Integral b => b -> a -> a

Repeat a value n times.

Given that this works on a Semigroup it is allowed to fail if you request 0 or fewer repetitions, and the default definition will do so.

By making this a member of the class, idempotent semigroups and monoids can upgrade this to execute in O(1) by picking stimes = stimesIdempotent or stimes = stimesIdempotentMonoid respectively.

stimesMonoid :: (Integral b, Monoid a) => b -> a -> a

This is a valid definition of stimes for a Monoid.

Unlike the default definition of stimes, it is defined for 0 and so it should be preferred where possible.

stimesIdempotent :: Integral b => b -> a -> a

This is a valid definition of stimes for an idempotent Semigroup.

When x <> x = x, this definition should be preferred, because it works in O(1) rather than O(log n).

stimesIdempotentMonoid :: (Integral b, Monoid a) => b -> a -> a

This is a valid definition of stimes for an idempotent Monoid.

When mappend x x = x, this definition should be preferred, because it works in O(1) rather than O(log n)

mtimesDefault :: (Integral b, Monoid a) => b -> a -> a

Repeat a value n times.

mtimesDefault n a = a <> a <> ... <> a  -- using <> (n-1) times

Implemented using stimes and mempty.

This is a suitable definition for an mtimes member of Monoid.

Semigroups

newtype Min a :: * -> *

Constructors

Min

Fields

getMin :: a

newtype Max a :: * -> *

Constructors

Max

Fields

getMax :: a

newtype First a :: * -> *

Use Option (First a) to get the behavior of First from Data.Monoid.

Constructors

First

Fields

getFirst :: a

newtype Last a :: * -> *

Use Option (Last a) to get the behavior of Last from Data.Monoid

Constructors

Last

Fields

getLast :: a

newtype WrappedMonoid m :: * -> *

Provide a Semigroup for an arbitrary Monoid.

Constructors

WrapMonoid

Fields

unwrapMonoid :: m

Re-exported monoids from Data.Monoid

newtype Dual a :: * -> *

The dual of a Monoid, obtained by swapping the arguments of mappend.

Constructors

Dual

Fields

getDual :: a

newtype Endo a :: * -> *

The monoid of endomorphisms under composition.

Constructors

Endo

Fields

appEndo :: a -> a

newtype All :: *

Boolean monoid under conjunction (&&).

Constructors

All

Fields

getAll :: Bool

newtype Any :: *

Boolean monoid under disjunction (||).

Constructors

Any

Fields

getAny :: Bool

newtype Sum a :: * -> *

Monoid under addition.

Constructors

Sum

Fields

getSum :: a

newtype Product a :: * -> *

Monoid under multiplication.

Constructors

Product

Fields

getProduct :: a

A better monoid for Maybe

newtype Option a :: * -> *

Option is effectively Maybe with a better instance of Monoid, built off of an underlying Semigroup instead of an underlying Monoid.

Ideally, this type would not exist at all and we would just fix the Monoid instance of Maybe

Constructors

Option

Fields

getOption :: Maybe a

option :: b -> (a -> b) -> Option a -> b

Fold an Option case-wise, just like maybe.

Difference lists of a semigroup

diff :: Semigroup m => m -> Endo m

This lets you use a difference list of a Semigroup as a Monoid.

cycle1 :: Semigroup m => m -> m

A generalization of cycle to an arbitrary Semigroup. May fail to terminate for some values in some semigroups.

ArgMin, ArgMax

data Arg a b :: * -> * -> *

Arg isn't itself a Semigroup in its own right, but it can be placed inside Min and Max to compute an arg min or arg max.

Constructors

Arg a b

type ArgMin a b = Min (Arg a b)

type ArgMax a b = Max (Arg a b)