Foreign.Marshal.Array.Compat

Documentation

callocArray :: Storable a => Int -> IO (Ptr a)

Like mallocArray, but allocated memory is filled with bytes of value zero.

callocArray0 :: Storable a => Int -> IO (Ptr a)

Like callocArray0, but allocated memory is filled with bytes of value zero.